Update DBC Structure

+++ WowPacketParser/DBC/Structures/BattleForAzeroth/DifficultyEntry.cs
@@ -10,7 +10,7 @@
         public byte FallbackDifficultyID;
         public byte MinPlayers;
         public byte MaxPlayers;
-        public byte Flags;
+        public ushort Flags;
         public byte ItemContext;
         public byte ToggleDifficultyID;
         public ushort GroupSizeHealthCurveID;         // unproven
